requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.12
python-dotenv==1.2.2
celery==5.6.3
redis==7.4.0
yfinance==1.3.0
pandas==3.0.2
numpy==2.4.4
scipy==1.17.1
requests==2.33.0
marshmallow==4.3.0
marshmallow-sqlalchemy==1.5.0